move js definitions to js-exports, definitions to separate project
