+++ backend/definitions/src/main/scala/aqua/definition/Definitions.scala
@@ -1,7 +1,8 @@
-package aqua.backend
+package aqua.definition
 
 import aqua.res.FuncRes
 import aqua.types.*
+import aqua.definition.*
 import io.circe.*
 import io.circe.parser.*
 import io.circe.syntax.*
