+++ cli/.js/src/main/scala/aqua/builder/ArgumentGetter.scala
@@ -6,6 +6,7 @@ import aqua.model.{LiteralModel, VarModel}
 import aqua.raw.ops
 import aqua.raw.ops.{Call, CallArrowRawTag}
 import aqua.raw.value.{LiteralRaw, VarRaw}
+import aqua.definition.*
 import cats.data.NonEmptyList
 
 import scala.concurrent.Promise
