+++ cli/.js/src/main/scala/aqua/builder/Finisher.scala
@@ -4,6 +4,7 @@ import aqua.backend.*
 import aqua.js.{CallJsFunction, FluencePeer, ServiceHandler}
 import aqua.model.{LiteralModel, VarModel}
 import aqua.raw.ops.{Call, CallArrowRawTag}
+import aqua.definition.*
 import aqua.raw.value.LiteralRaw
 import cats.data.NonEmptyList
