+++ cli/.js/src/main/scala/aqua/builder/Service.scala
@@ -2,6 +2,7 @@ package aqua.builder
 
 import aqua.backend.*
 import aqua.js.{CallJsFunction, FluencePeer, ServiceHandler}
+import aqua.definition.*
 import cats.data.NonEmptyList
 import scribe.Logging
